OOP changes
